Refactor unit tests to conform to existing test patterns
